Add basic authorization
- Only the author of campground can now edit and delete the campground
- Only the author of reviews can now delete the reviews

+++ routes/campgrounds.js
@@ -1,19 +1,8 @@
 const catchAsync = require('../utils/catchAsync');
 const express = require('express');
-const ExpressError = require('../utils/ExpressError');
 const router = express.Router();
 const Campground = require('../models/campground')
-const { campgroundSchema } = require('../schemas.js');
-const { isLoggedIn, storeReturnTo} = require('../middleware');
-
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    }
-    next();
-}
+const { isLoggedIn, validateCampground, isAuthor } = require('../middleware')
 
 router.get('/', async (req, res) => {
     const campgrounds = await Campground.find({});
@@ -28,6 +17,7 @@ router.get('/new', isLoggedIn, (req, res) => {
 router.post('/', isLoggedIn, validateCampground, catchAsync(
     async (req, res) => {
         const campground = new Campground(req.body.campground);
+        campground.author = req.user._id; // req.user is added by passport
         await campground.save();
         req.flash('success', 'Successfully made a new campground!');
         res.redirect(`/campgrounds/${campground._id}`);
@@ -36,7 +26,12 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(
 
 router.get('/:id', catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findById(id).populate('reviews');
+    const campground = await Campground.findById(id).populate({
+        // populate reviews and their authors
+        path: 'reviews',
+        populate: { path: 'author' }
+    }).populate('author'); // it's the `author` field in campgroundSchema
+    console.log(campground);
     if (!campground) {
         req.flash('error', 'Cannot find that campground!');
         return res.redirect('/campgrounds'); // TODO: why return?
@@ -44,7 +39,7 @@ router.get('/:id', catchAsync(async (req, res) => {
     res.render('campgrounds/show', { campground });
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => { // TODO: bug
     const { id } = req.params;
     const campground = await Campground.findById(id);
     if (!campground) {
@@ -54,8 +49,13 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
     res.render('campgrounds/edit', { campground });
 }))
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', "You do not have permission to do that!");
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndUpdate(id, req.body.campground);
     req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${id}`);
@@ -63,6 +63,11 @@ router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) =
 
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', "You do not have permission to do that!");
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground');
     res.redirect('/campgrounds');
